bump teaclave-sdk to v1.1.6-testing #1074 (#1076)
* update rust sgx sdk to v1.1.6-testing, add exact commit hash to script

* Makefile: update pin-sgx

* update toolchain

* update teaclave to 1.1.6-testing

* enclave: update teaclave version

* itp-top-pool: remove attribute `vec_retain_mut` as it has been stable since 1.61.0

* itp-attestation-handler: remove const_err as it is a hard error in Rust now

* clippy fixes

* allow clippy::result_large_err

* fix clippy in enclave-runtime

* fix clippy: feature EVM

* fix clippy Teeracle feature

* clippy: allow or_fun() in fn pruned() as .defaul() is not available in enclave-runtime

* clippy: fix allowed large_enum_variant lint in tls-websocket-server

* clippy: fix large_enum_variant in app-libs/stf

* fix mio and net2 conflicts

* fix needless borrow

* clippy: unused_io_amount

* add explicit revision for sgx_crypto_helper

* Update teaclave explicitly in stf-state-handler

* stf-state-handler: add explicit feature to sgx_crypto_helper

* fixup! fix mio and net2 conflicts

* make sure all teaclave dependency is on the same version

+++ core-primitives/enclave-api/src/enclave_base.rs
@@ -185,7 +185,7 @@ impl EnclaveBase for Enclave {
 		let mut retval = sgx_status_t::SGX_SUCCESS;
 
 		let pubkey_size = SHIELDING_KEY_SIZE;
-		let mut pubkey = vec![0u8; pubkey_size as usize];
+		let mut pubkey = vec![0u8; pubkey_size];
 
 		let result = unsafe {
 			ffi::get_rsa_encryption_pubkey(
@@ -207,7 +207,7 @@ impl EnclaveBase for Enclave {
 
 	fn get_ecc_signing_pubkey(&self) -> EnclaveResult<ed25519::Public> {
 		let mut retval = sgx_status_t::SGX_SUCCESS;
-		let mut pubkey = [0u8; SIGNING_KEY_SIZE as usize];
+		let mut pubkey = [0u8; SIGNING_KEY_SIZE];
 
 		let result = unsafe {
 			ffi::get_ecc_signing_pubkey(
@@ -226,7 +226,7 @@ impl EnclaveBase for Enclave {
 
 	fn get_mrenclave(&self) -> EnclaveResult<[u8; MR_ENCLAVE_SIZE]> {
 		let mut retval = sgx_status_t::SGX_SUCCESS;
-		let mut mr_enclave = [0u8; MR_ENCLAVE_SIZE as usize];
+		let mut mr_enclave = [0u8; MR_ENCLAVE_SIZE];
 
 		let result = unsafe {
 			ffi::get_mrenclave(
@@ -251,7 +251,7 @@ fn init_parentchain_components_ffi(
 	let mut retval = sgx_status_t::SGX_SUCCESS;
 
 	let latest_header_size = HEADER_MAX_SIZE;
-	let mut latest_header = vec![0u8; latest_header_size as usize];
+	let mut latest_header = vec![0u8; latest_header_size];
 
 	let result = unsafe {
 		ffi::init_parentchain_components(
